Move ErrorContains to an internal package.
