+++ builder/remotecontext/remote_test.go
@@ -10,8 +10,8 @@ import (
 	"testing"
 
 	"github.com/docker/docker/builder"
+	"github.com/docker/docker/internal/testutil"
 	"github.com/docker/docker/pkg/archive"
-	"github.com/docker/docker/pkg/testutil"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
