+++ daemon/config/config_test.go
@@ -8,8 +8,8 @@ import (
 	"testing"
 
 	"github.com/docker/docker/daemon/discovery"
+	"github.com/docker/docker/internal/testutil"
 	"github.com/docker/docker/opts"
-	"github.com/docker/docker/pkg/testutil"
 	"github.com/spf13/pflag"
 	"github.com/stretchr/testify/assert"
 )
