+++ image/store_test.go
@@ -4,8 +4,8 @@ import (
 	"runtime"
 	"testing"
 
+	"github.com/docker/docker/internal/testutil"
 	"github.com/docker/docker/layer"
-	"github.com/docker/docker/pkg/testutil"
 	"github.com/opencontainers/go-digest"
 	"github.com/stretchr/testify/assert"
 )
