+++ integration/container/create_test.go
@@ -8,7 +8,7 @@ import (
 	"github.com/docker/docker/api/types/container"
 	"github.com/docker/docker/api/types/network"
 	"github.com/docker/docker/integration/util/request"
-	"github.com/docker/docker/pkg/testutil"
+	"github.com/docker/docker/internal/testutil"
 )
 
 func TestCreateFailsWhenIdentifierDoesNotExist(t *testing.T) {
